test: break out shared test utils.
