+++ tests/serverpod_test_server/test_e2e/cloud_storage_test.dart
@@ -2,9 +2,8 @@ import 'dart:typed_data';
 
 import 'package:http/http.dart' as http;
 import 'package:serverpod_test_client/serverpod_test_client.dart';
+import 'package:serverpod_test_server/test_util/config.dart';
 import 'package:test/test.dart';
-
-import 'config.dart';
 
 ByteData createByteData(int len) {
   var ints = Uint8List(len);
