+++ tests/serverpod_test_server/test_e2e/column_operations_legacy/column_enum_test.dart
@@ -1,7 +1,6 @@
 import 'package:serverpod_test_client/serverpod_test_client.dart';
+import 'package:serverpod_test_server/test_util/config.dart';
 import 'package:test/test.dart';
-
-import '../config.dart';
 
 Future<void> _createTestDatabase(Client client) async {
   var firstEnum = Types(anEnum: TestEnum.one);
