+++ tests/serverpod_test_server/test_e2e/column_operations_legacy/column_uuid_test.dart
@@ -1,7 +1,6 @@
 import 'package:serverpod_test_client/serverpod_test_client.dart';
+import 'package:serverpod_test_server/test_util/config.dart';
 import 'package:test/test.dart';
-
-import '../config.dart';
 
 final firstUuid = UuidValue('6948DF80-14BD-4E04-8842-7668D9C001F5');
 final secondUuid = UuidValue('4B8302DA-21AD-401F-AF45-1DFD956B80B5');
