+++ tests/serverpod_test_server/test_e2e/connection_test.dart
@@ -2,10 +2,9 @@ import 'dart:typed_data';
 
 import 'package:serverpod_test_client/serverpod_test_client.dart';
 import 'package:serverpod_test_client/src/custom_classes.dart';
+import 'package:serverpod_test_server/test_util/config.dart';
 import 'package:serverpod_test_shared/serverpod_test_shared.dart';
 import 'package:test/test.dart';
-
-import 'config.dart';
 
 Future<void> setupTestData(Client client) async {
   await client.basicDatabaseLegacy.deleteAllSimpleTestData();
