+++ tests/serverpod_test_server/test_e2e/exception_test.dart
@@ -1,9 +1,8 @@
 import 'package:serverpod_test_client/serverpod_test_client.dart'
     as test_client;
 import 'package:serverpod_test_client/serverpod_test_client.dart';
+import 'package:serverpod_test_server/test_util/config.dart';
 import 'package:test/test.dart';
-
-import 'config.dart';
 
 void main() {
   var client = test_client.Client(
