+++ tests/serverpod_test_server/test_e2e/redis_test.dart
@@ -1,9 +1,8 @@
 import 'dart:async';
 
 import 'package:serverpod_test_client/serverpod_test_client.dart';
+import 'package:serverpod_test_server/test_util/config.dart';
 import 'package:test/test.dart';
-
-import 'config.dart';
 
 void main() {
   var client = Client(
